requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
fpdf2
pdfplumberscikit-learn==1.8.0
